refactor: move github module to lib/zero/ and fix schema violation
Part of the lib/ restructuring epic (#7697). Moves GitHub App webhook
handlers (issues, installations) and auth utilities into the zero/ layer.

Also fixes a schema violation: replaces direct zero-agent schema import
in issue-event.ts with a new resolveAgentId() function in
zero-compose-service, keeping the github module's DB access properly
encapsulated.

Closes #7733
