+++ turbo/apps/web/src/lib/zero/github/handlers/installation-event.ts
@@ -1,10 +1,10 @@
 import { z } from "zod";
 import { eq, and } from "drizzle-orm";
-import { githubInstallations } from "../../../db/schema/github-installation";
+import { githubInstallations } from "../../../../db/schema/github-installation";
 import { getInstallationAccessToken } from "../github-app";
-import { encryptSecretValue } from "../../shared/crypto/secrets-encryption";
-import { env } from "../../../env";
-import { logger } from "../../logger";
+import { encryptSecretValue } from "../../../shared/crypto/secrets-encryption";
+import { env } from "../../../../env";
+import { logger } from "../../../logger";
 
 const log = logger("github:installation-event");
